supervised: cast states to bf16 in collate to halve H2D payload
Move the existing bf16 cast earlier into _trajectory_collate_fn so the
H2D staging copy (CPU-bound under WSL2 pin_memory=False) carries half the
bytes; drop the now-redundant fp32 up-casts in train.py, evaluate(), and
analyze() since autocast handles dtype consistency inside the forward pass.

+++ src/elitefurretai/supervised/utils.py
@@ -375,7 +375,7 @@ def evaluate(
     for batch in _eval_iterator:
         if device != "cuda":
             batch = {k: v.to(device) for k, v in batch.items()}
-        states = batch["states"].to(torch.float32)
+        states = batch["states"]
         actions = batch["actions"]
         action_masks = batch["action_masks"] if "action_masks" in batch else None
         masks = batch["masks"] if "masks" in batch else None
@@ -862,7 +862,7 @@ def analyze(
     for batch in _analyze_iterator:
         if device != "cuda":
             batch = {k: v.to(device) for k, v in batch.items()}
-        states = batch["states"].to(torch.float32)
+        states = batch["states"]
         if state_input_dim is not None and state_input_dim < states.shape[-1]:
             states = states[..., :state_input_dim]
         actions = batch["actions"]
